Add support for building infinite projection matrices. Both standard and reverse depth.
The infinite reverse projection matrix is the only one we use, and it is the one with the best precision and the simplest construction. So really the one most everyone should use by default.

Probably should rename the perspective_glu function here also, or maybe (depending on how opionated we want to be) simply remove it? It is only useful in OpenGL due to the -1,1 depth range and it should have a `_rh` suffix to indicate it is right-handed.