Removed the dependency on the `approx` crate.
This was primiarily used for unit tests. It has been replaced by a
simple `assert_approx_eq` macro and a `FloatCompare` trait defined in
the test `support` module which is implemented for all `glam` types.

Each `glam` type has had an `abs_diff_eq` method added which returns
true if the absolute difference of each element is less than or equal to
a given `max_abs_diff` value.

This is mostly sufficient for the purpose of `glam`'s tests.

Fixes #6.