Trying different ways of benching.
Getting rid of the vectors of input data seemed to make a lot of
benchmarks take 300ps, which suggested they weren't doing anything.
Changing to `iter_batch_ref` seemed to make a lot of them a lot slower.

Ultimately I'm not sure what's going on. It would be good to get actual
timing for a function but at least the benchmarks prior to this change
seemed to be consistent relative to each other.

Would need to disassmble what is actually getting measured by Criterion
to know for sure. Ideally benchmarks aren't getting cache misses and
aren't get optimised away by the compiler.