Migrated from using cfg-if to setting mutually exclusive cfg values via build.rs (#55)
I was having a few issues with cfg-if like rustfmt didn't seem to like it and neither did tarpaulin.

Separate to that it seemed like checking target features and features in rust code was getting complicated with more targets and features. So instead I'm doing the more complex logic in rust code to generate mutually exclusive rust-cfg's that can be checked in Rust code. This is independent of whether or not cfg-if it being used.

For example:

```
    pub fn length(self) -> f32 {
        #[cfg(vec3sse2)]
        unsafe {
            _mm_cvtss_f32(_mm_sqrt_ss(self.dot_as_m128(self)))
        }

        #[cfg(vec3f32)]
        {
            self.dot(self).sqrt()
        }
    }
```

`cfg-if` using `else if` and `else` blocks but we don't have that with plain on `#[cfg(name)]`, instead I'm creating mutually exclusive cfgs which can both exist in the same function because they'll never both be compiled.