Split `Vec3` into `__m128` and `f32` versions.
`Vec3` is now always stored as `(f32, f32, f32)` and always has a natural
alignment of 4 bytes.

`Vec3Align16` is stored as `__m128` when available and is always 16 byte
aligned.

`Vec3Mask` and `Vec3MaskAlign16` correspond to the `Vec3` and
`Vec3Align16` types.

`Mat3` uses the `Vec3` type for storage and now always has a natual
alignment of 4 bytes.

Currently the `Vec3Align16` and `Vec3MaskAlign16` types are not tested,
these will need to be added.