Rename types containing Vec3A with suffix 3A.
The intention is this is clearer than these types contain `Vec3A` and
thus are 16 byte aligned and have internal padding.

Also there is typically a significant performance advantage to using
`Vec3A` on platforms with SSE2 support.

There is still a bit of a mix of `Vec3` and `Vec3A` parameters.
Generally for methods than may be called a lot like transforms there are
methods that take both.