Renamed `Affine3` to `Affine3A`, `TransformRT` to `Isometry3A` and `TransformSRT` to `Transform3A` (#174)
For the most part this renaming is to indicate that these types are `Vec3A` based which implies they are 16 byte aligned and contain internal padding.

Renamed transform types because these names are used in other Rust crates like nalgebra and ultraviolet and they work a bit better with the `A` suffix.

All transform types are currently on the `transform-types` feature. I haven't decided yet if they should be part of the main library, a separate crate or on a feature.

Added benches for `Mat3A`.

Removed the `to_simd` / `from_simd` code from Mat3, it was slower.