Added compile time asserts to check size and alignment.
Size and alignment should be consistent for types using SIMD even if
SIMD is not available. Alignment is only not enforced if `scalar-math`
is used. Because there was no build that didn't have SIMD support some
inconsistent alignment slipped through. It is the intention to support
SIMD on all stable architectures eventually which is why size and
alignment are intended to be consistent unless SIMD is disabled
entirely.

Also did some work on module docs.