Remove legacy `::from()` truncation mention in docs; add more intradoc links
`::from()` to truncate `VecN` to `Vec(N - 1)` has been removed long ago
in #195, and should not be mentioned anymore in the docs.  Also replace
more type and function references with proper intradocs, making it
easier for the reader to click through to the thing that is being
mentioned as well as validate such documentation (in particular:
validity of intradoc links) on the CI in the future.